The most significant changes in the code are related to type casting, the use of constant expressions, order of member variable declarations, return of const values, addition of default values to member variables, and the replacement of raw pointers with smart pointers. There were also changes in file names to follow a naming convention, removal of self-inclusion, and an update in the high scores list.
1. Type Casting: The developer has ensured type safety by replacing direct assignments or operations with their type-casted versions. For example, integer values are cast to float where a float is expected, and vice versa (Reference: Code Changes).

2. Constant Expressions: Magic numbers have been replaced with more descriptive constant expressions for better readability and maintainability (Reference: Code Changes).

3. Member Variable Declarations: The order of member variable declarations in classes has been changed to match the order in which they are initialized in the constructor's initializer list (Reference: Code Changes).

4. Return of Const Values: Some functions have been updated to return const values, ensuring that the returned value cannot be modified (Reference: Code Changes).

5. Default Values: Default values have been added to member variables in their declarations (Reference: Code Changes).

6. Smart Pointers: In the file `shipCannonHandler_PG.cpp`, raw pointers have been replaced with smart pointers for better memory management (Reference: `shipCannonHandler_PG.cpp`).

7. File Names: The file names for `shipCannonHandler_PG.cpp` and `shipCannonHandler_PG.h` were changed to `ShipCannonHandler_PG.cpp` and `ShipCannonHandler_PG.h` respectively (Reference: SFML Project 1.vcxproj).

8. Self-Inclusion: In the file `Cannonball_PG.h`, the developer has removed the self-inclusion of `Cannonball_PG.h` (Reference: `Cannonball_PG.h`).

10. File Removal: The file `Cannonball_PG.cpp` has been removed, because it was a dupe.

+++ AgarGame/Resource/Debugger_AG.cpp
@@ -41,9 +41,9 @@ void Debugger::drawVector(const sf::Vector2f& start, const sf::Vector2f& directi
 
 void Debugger::drawCell(sf::RenderWindow& window, int x, int y, Map& map, sf::Color color) {
     sf::RectangleShape rectangle;
-    rectangle.setSize(sf::Vector2f(map.getCellSize(), map.getCellSize()));
+    rectangle.setSize(sf::Vector2f(static_cast<float>(map.getCellSize()), static_cast<float>(map.getCellSize())));
     rectangle.setFillColor(color);
-    rectangle.setPosition(x * map.getCellSize(), y * map.getCellSize());
+    rectangle.setPosition(static_cast<float>(x * map.getCellSize()), static_cast<float>(y * map.getCellSize()));
     window.draw(rectangle);
 
     //std::cout << "Drawing cell\n";
